+++ packages/common/directives/src/public_api.ts
@@ -1,2 +1,12 @@
+export { AmountFormatDirective } from './amount-format/amount-format.directive';
+export { AmountFormatModule } from './amount-format/amount-format.module';
+export { ConvertCaseDirective } from './convert-case/convert-case.directive';
+export { ConvertCaseModule } from './convert-case/convert-case.module';
+export { DisableControlDirective } from './disable-control/disable-control.directive';
+export { DisableControlModule } from './disable-control/disable-control.module';
 export { InitialFocusDirective } from './initial-focus/initial-focus.directive';
 export { InitialFocusModule } from './initial-focus/initial-focus.module';
+export { MaxLengthDirective } from './max-length/max-length.directive';
+export { MaxLengthModule } from './max-length/max-length.module';
+export { TrimInputDirective } from './trim-input/trim-input.directive';
+export { TrimInputModule } from './trim-input/trim-input.module';