+++ packages/common/validators/src/public_api.ts
@@ -1,0 +1,1 @@
+export { emptyValidator, EmptyValidatorFn } from './empty.validator';