+++ packages/ng-table-builder/integration/app/src/samples/sample-fourteen/not-found.component.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
+
+@Component({
+    selector: 'not-found',
+    template: `<b>Not found</b>`,
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    encapsulation: ViewEncapsulation.None
+})
+export class NotFoundComponent {}